# openapi-schema.yaml
openapi: 3.1.0
info:
  title: WHOOP GPT API
  description: API for retrieving WHOOP health data
  version: 1.0.0
servers:
  - url: https://whoopgpt-production.up.railway.app
    description: Railway deployment
paths:
  /api/recovery:
    get:
      summary: Get latest WHOOP recovery for the user
      operationId: getRecovery
      parameters:
        - name: token_id
          in: query
          description: Authentication token ID
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Recovery data
          content:
            application/json:
              schema:
                type: object
                properties:
                  date:
                    type: string
                    description: Date of the recovery score
                  recovery_score:
                    type: integer
                    description: Overall recovery score (0-100)
                  hrv:
                    type: number
                    description: Heart Rate Variability in milliseconds
                  rhr:
                    type: number
                    description: Resting Heart Rate in beats per minute
                  sleep_quality:
                    type: number
                    description: Sleep quality score (0-100)
                  user_status:
                    type: string
                    description: Status of the user
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  auth_required:
                    type: boolean
                  auth_url:
                    type: string
  /api/sleep:
    get:
      summary: Get latest WHOOP sleep data for the user
      operationId: getSleep
      parameters:
        - name: token_id
          in: query
          description: Authentication token ID
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Sleep data
          content:
            application/json:
              schema:
                type: object
                properties:
                  date:
                    type: string
                    description: Date of the sleep record
                  score:
                    type: integer
                    description: Overall sleep score (0-100)
                  total_duration_minutes:
                    type: integer
                    description: Total sleep duration in minutes
                  efficiency:
                    type: number
                    description: Sleep efficiency percentage
                  disturbances:
                    type: integer
                    description: Number of sleep disturbances
                  deep_sleep_minutes:
                    type: integer
                    description: Time spent in deep sleep (minutes)
                  rem_sleep_minutes:
                    type: integer
                    description: Time spent in REM sleep (minutes)
                  light_sleep_minutes:
                    type: integer
                    description: Time spent in light sleep (minutes)
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  auth_required:
                    type: boolean
                  auth_url:
                    type: string
  /api/profile:
    get:
      summary: Get WHOOP user profile
      operationId: getProfile
      parameters:
        - name: token_id
          in: query
          description: Authentication token ID
          required: false
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  first_name:
                    type: string
                    description: User's first name
                  last_name:
                    type: string
                    description: User's last name
                  email:
                    type: string
                    description: User's email address
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  auth_required:
                    type: boolean
                  auth_url:
                    type: string
  /api/recovery/history:
    get:
      summary: Get historical WHOOP recovery data for a specified time period
      operationId: getRecoveryHistory
      parameters:
        - name: token_id
          in: query
          description: Authentication token ID
          required: false
          schema:
            type: string
        - name: days
          in: query
          description: Number of days to get history for (1 to 180, default 30)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 180
            default: 30
      responses:
        "200":
          description: Historical recovery data
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: object
                    description: Summary statistics for the time period
                    properties:
                      total_records:
                        type: integer
                        description: Total number of records in the time period
                      days_requested:
                        type: integer
                        description: Number of days requested for the query
                      avg_recovery_score:
                        type: integer
                        description: Average recovery score over the time period
                      avg_hrv:
                        type: number
                        description: Average HRV over the time period
                      avg_rhr:
                        type: integer
                        description: Average resting heart rate over the time period
                      highest_recovery_score:
                        type: integer
                        description: Highest recovery score in the time period
                      lowest_recovery_score:
                        type: integer
                        description: Lowest recovery score in the time period
                      data_by_week:
                        type: object
                        description: Data organized by week
                        additionalProperties:
                          type: object
                          properties:
                            avg_recovery_score:
                              type: integer
                              description: Average recovery score for the week
                            avg_hrv:
                              type: number
                              description: Average HRV for the week
                            avg_rhr:
                              type: integer
                              description: Average resting heart rate for the week
                  records:
                    type: array
                    description: All individual recovery records
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          description: Date of the recovery score
                        recovery_score:
                          type: integer
                          description: Overall recovery score (0-100)
                        hrv:
                          type: number
                          description: Heart Rate Variability in milliseconds
                        rhr:
                          type: number
                          description: Resting Heart Rate in beats per minute
                        user_status:
                          type: string
                          description: Status of the user
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  auth_required:
                    type: boolean
                  auth_url:
                    type: string
  /auth:
    get:
      summary: Initiate WHOOP OAuth flow
      operationId: startAuth
      responses:
        "302":
          description: Redirect to WHOOP authorization page
  /logout:
    get:
      summary: Logout and clear session
      operationId: logout
      parameters:
        - name: token_id
          in: query
          description: Authentication token ID
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Successfully logged out
          content:
            text/plain:
              schema:
                type: string